refactor(protocol): align typed RPC dispatch with current boundaries
