+++ packages/daemon/src/protocol/daemon-protocol.contract.ts
@@ -696,7 +696,6 @@ export {
   DaemonProtocolContractError,
   isDaemonGuiActionMethod,
   isDaemonGuiReadMethod,
-  isDaemonRpcMethod,
   isDaemonStreamMethod,
   parseDaemonRpcParams,
   serializeDaemonRpcCall,
